+++ HISTORY.rst
@@ -4,6 +4,7 @@ History
 0.1.8 (unreleased)
 ------------------
 
+- Sorting for all columns: #16
 - Login to 3Di from Rana using a personal API key: #15